Split LEDMatrxGFX and LEDStripGFX
