+++ include/effects/effectmanager.h
@@ -42,7 +42,7 @@
 #include <vector>
 #include <math.h>
 #include "colorutils.h"
-#include "ledmatrixgfx.h"
+#include "ledstripgfx.h"
 #include "ledstripeffect.h"
 #include "globals.h"
 #include "misceffects.h"
@@ -76,7 +76,7 @@ class EffectManager
 	bool             _bPlayAll;
 
 	std::unique_ptr<bool []> _abEffectEnabled;
-	std::shared_ptr<LEDMatrixGFX> * _gfx;
+	std::shared_ptr<LEDStripGFX> * _gfx;
 	std::unique_ptr<LEDStripEffect> _pRemoteEffect;
 
 public:
@@ -84,7 +84,7 @@ public:
 	static const uint csFadeButtonSpeed = 15 * 1000;
 	static const uint csSmoothButtonSpeed = 60 * 1000;
 
-	EffectManager(LEDStripEffect ** pEffects, size_t cEffects, std::shared_ptr<LEDMatrixGFX> * gfx)
+	EffectManager(LEDStripEffect ** pEffects, size_t cEffects, std::shared_ptr<LEDStripGFX> * gfx)
 		  : _ppEffects(pEffects),
 	  	    _cEffects(cEffects),
 			_cEnabled(0),
@@ -106,7 +106,7 @@ public:
 		ClearRemoteColor();
 	}
 
-	std::shared_ptr<LEDMatrixGFX> operator [](size_t index)
+	std::shared_ptr<LEDStripGFX> operator [](size_t index)
 	{
 		return _gfx[index];
 	}
