+++ include/effects/meteoreffect.h
@@ -69,7 +69,7 @@ public:
 	
 	}
 
-	virtual void Init(std::shared_ptr<LEDMatrixGFX> pGFX, size_t meteors = 4, uint size = 4, uint decay = 3, double minSpeed = 0.5, double maxSpeed = 0.5)
+	virtual void Init(std::shared_ptr<LEDStripGFX> pGFX, size_t meteors = 4, uint size = 4, uint decay = 3, double minSpeed = 0.5, double maxSpeed = 0.5)
 	{
 		meteorCount = meteors;
 		meteorSize = size;
@@ -102,7 +102,7 @@ public:
 		bLeft[iMeteor] = !bLeft[iMeteor];
 	}
 
-	virtual void Draw(std::shared_ptr<LEDMatrixGFX> pGFX)
+	virtual void Draw(std::shared_ptr<LEDStripGFX> pGFX)
 	{
 		static CHSV hsv;
 		hsv.val = 255;
@@ -167,7 +167,7 @@ class MeteorEffect : public LEDStripEffect
 {
   private:
 	MeteorChannel   _Meteors[NUM_CHANNELS];
-    std::shared_ptr<LEDMatrixGFX> * _gfx;
+    std::shared_ptr<LEDStripGFX> * _gfx;
 
 	int				_cMeteors;
 	uint8_t         _meteorSize;
@@ -186,7 +186,7 @@ class MeteorEffect : public LEDStripEffect
 		_meteorSpeedMax = maxSpeed;
     }
 
-    virtual bool Init(std::shared_ptr<LEDMatrixGFX> gfx[NUM_CHANNELS])	
+    virtual bool Init(std::shared_ptr<LEDStripGFX> gfx[NUM_CHANNELS])	
     {
         _gfx = gfx;
         if (!LEDStripEffect::Init(gfx))
