+++ include/effects/vueffect.h
@@ -63,7 +63,7 @@ public:
         _colorSpeed = colorSpeed;
 	}
 
-    virtual bool Init(std::shared_ptr<LEDMatrixGFX> gfx[NUM_CHANNELS])
+    virtual bool Init(std::shared_ptr<LEDStripGFX> gfx[NUM_CHANNELS])
     {
         if (!LEDStripEffect::Init(gfx))
             return false;
