+++ include/globals.h
@@ -407,8 +407,6 @@ extern RemoteDebug Debug;           // Let everyone in the project know about it
     #define ENABLE_REMOTE           1   // IR Remote Control
     #define ENABLE_AUDIO            1   // Listen for audio from the microphone and process it
 
-    #define USE_LCD                 1
-    
     #define DEFAULT_EFFECT_INTERVAL     (60*60*24)
 
     #if USE_PSRAM
@@ -876,6 +874,11 @@ extern RemoteDebug Debug;           // Let everyone in the project know about it
     #define NUM_INFO_PAGES 1
 
     #define COLOR_ORDER EOrder::RGB
+#endif
+
+#ifdef MESMERIZER
+#include "MatrixHardware_ESP32_Custom.h"
+#include <SmartMatrix.h>
 #endif
