+++ platformio.ini
@@ -31,6 +31,7 @@ lib_deps        = crankyoldgit/IRremoteESP8266  @ ^2.7.20
                   
                   
 ; Libs needed to link with an M5StickC
+
 [m5stick-c]
 lib_deps        = ${common.lib_deps}
                   m5stack/M5StickC     @ ^0.2.3 
@@ -192,19 +193,19 @@ board         = esp-wrover-kit
 upload_speed  = 1500000
 monitor_speed = 115200
 upload_port   = COM12
-monitor_port   = COM12
+monitor_port  = COM12
 build_flags   = -DMESMERIZER=1
                 -DUSE_PSRAM=1
                 -DBOARD_HAS_PSRAM=1
                 -mfix-esp32-psram-cache-issue
-                -DUSE_SCREEN=1
-                -DUSE_LCD=1
                 -DEEBUG=1
                 -D_IR_ENABLE_DEFAULT_=false  ; don't automatically include every remote control decoder
                 -DDECODE_NEC=true            ; enable whichever you need for your remote.  Try not disabling above to figure out which it is.
                 -std=gnu++17
                 -Ofast 
-lib_deps      = ${m5stick-c-plus.lib_deps}
+                
+lib_deps      =  ${common.lib_deps}
+                 SmartMatrix
 
 [env:spectrum]
 board         = m5stick-c                    ; Requires the M5 Stick C Plus (note the Plus)
